backend/apps/triage/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import TriageViewSet

router = DefaultRouter()
router.register("", TriageViewSet, basename="triage")

urlpatterns = [path("", include(router.urls))]
